Use more specific services to stop/start during cert creation process
This is useful when you need to stop different services depending
on the certificates. For example, if you have a first certificate
that uses varnish and a second one using nginx you can have a
configuration that looks like this:

certbot_certs:
  - email: janedoe@example.com
    stop_services:
      - varnish
    domains:
      - example1.com
  - email: janedoe@example.com
    stop_services:
      - nginx
    domains:
      - example2.com

This way, you don't have to stop both varnish and nginx during cert
creation process.